Initial implementation, continued (43):
- tdsl_command_context/execute_query: now returns query_result instead
  of rows_affected
  * query_result object reflects statement execution status obtained from
    the DONE token
  * also contains rows_affected
- tdsl_command_context/execute_rpc: added float4/float8 support
  * moved variable type to fixed type conversion logic to var_to_fixed()
    function
- tdsl_macro: added TDSL_CANNOT_HAPPEN macro
- tdsl_field: implemented sql_smalldatetime, sql_datetime
  * ... and added unit tests.
- tdsl_field: as_impl enable_when::integral requirement relaxed to
  enable_when::arithmetic
- type_traits: added enable_when::arithmetic

+++ tdslite/include/tdslite/detail/tdsl_packet_handler_result.hpp
@@ -17,7 +17,7 @@
 namespace tdsl {
     template <typename Status, Status DefaultStatusV>
     struct packet_handler_result {
-        Status status               = {DefaultStatusV};
+        Status status             = {DefaultStatusV};
         /**
          * `needed_bytes` is a way to signal network layer
          * to know at least how many bytes more are needed
@@ -30,7 +30,7 @@ namespace tdsl {
          * implementation has to support it via constructs such as
          * read(transfer_at_least{needed_bytes}).
          */
-        tdsl::uint32_t needed_bytes = {0};
+        tdsl::size_t needed_bytes = {0};
     };
 } // namespace tdsl
